Global sum reorganization
