--- Include/Core/hspinors.h
@@ -1,14 +1,0 @@
-// TODO: Move this into suN.h
-
-#ifndef HSPINORS_H
-#define HSPINORS_H
-
-typedef struct _suNf_hspinor {
-  suNf_vector c[2];
-} suNf_hspinor;
-
-typedef struct _suNf_spinor_flt {
-    suNf_vector_flt c[2];
-} suNf_hspinor_flt;
-
-#endif
