+++ Include/Geometry/TMPL/strided_reads_gpu.h.tmpl
@@ -6,19 +6,19 @@
 #include "Utils/generics.h"
 
 #if !defined _FIELD_NAME
-#error Missing _FIELD_NAME in strided_reads.cu
+#error Missing _FIELD_NAME in strided_reads_gpu.h
 #endif
 #if !defined _FIELD_TYPE
-#error Missing _FIELD_TYPE in strided_reads.cu
+#error Missing _FIELD_TYPE in strided_reads_gpu.h
 #endif
 #if !defined _SITE_TYPE
-#error Missing _SITE_TYPE in strided_reads.cu
+#error Missing _SITE_TYPE in strided_reads_gpu.h
 #endif
 #if !defined _FIELD_DIM
-#error Missing _FIELD_DIM in strided_reads.cu
+#error Missing _FIELD_DIM in strided_reads_gpu.h
 #endif
 #if !defined _REAL
-#error Missing _REAL in strided_reads.cu
+#error Missing _REAL in strided_reads_gpu.h
 #endif
 
 _GPU_FUNC(void, read_, _FIELD_NAME, (_SITE_TYPE *s, _FIELD_TYPE *in, int ix, int comp));
