--- Include/Geometry/direction.h
@@ -1,9 +1,0 @@
-#ifndef DIRECTION_H
-#define DIRECTION_H
-
- enum DIRECTION {
-      UP = 0,
-      DOWN = 1
-};
-
-#endif
