+++ Include/Inverters/TMPL/linear_algebra.h.tmpl
@@ -37,96 +37,97 @@
 #endif
 
 
+
 /* s1=s2 */
-_DECLARE(void, spinor_field_copy, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_copy, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* Re <s1,s2> */
-_DECLARE(double, spinor_field_prod_re, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(double, spinor_field_prod_re, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* Im <s1,s2> */
-_DECLARE(double, spinor_field_prod_im, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(double, spinor_field_prod_im, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* <s1,s2> */
-_DECLARE(hr_complex, spinor_field_prod, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(hr_complex, spinor_field_prod, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* Re <g5*s1,s2> */
-_DECLARE(double, spinor_field_g5_prod_re, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(double, spinor_field_g5_prod_re, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* Im <g5*s1,s2> */
-_DECLARE(double, spinor_field_g5_prod_im, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(double, spinor_field_g5_prod_im, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* Re <s1,s1> */
-_DECLARE(double, spinor_field_sqnorm, (_SPINOR_FIELD_TYPE*));
+_DECLARE(double, spinor_field_sqnorm, (_SPINOR_FIELD_TYPE *s1));
 
 /* s1+=r*s2 r real */
-_DECLARE(void, spinor_field_mul_add_assign, (_SPINOR_FIELD_TYPE*, _REAL r, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_mul_add_assign, (_SPINOR_FIELD_TYPE *s1, _REAL r, _SPINOR_FIELD_TYPE *s2));
 
 /* s1+=c*s2 c complex */
-_DECLARE(void, spinor_field_mulc_add_assign, (_SPINOR_FIELD_TYPE*, _COMPLEX c, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_mulc_add_assign, (_SPINOR_FIELD_TYPE *s1, _COMPLEX c, _SPINOR_FIELD_TYPE *s2));
 
 /* s1+=c*g5*s2 c complex */
-_DECLARE(void, spinor_field_g5_mulc_add_assign, (_SPINOR_FIELD_TYPE*, _COMPLEX c, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_g5_mulc_add_assign, (_SPINOR_FIELD_TYPE *s1, _COMPLEX c, _SPINOR_FIELD_TYPE *s2));
 
 /* s1=r*s2 */
-_DECLARE(void, spinor_field_mul, (_SPINOR_FIELD_TYPE*, _REAL r, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_mul, (_SPINOR_FIELD_TYPE *s1, _REAL r, _SPINOR_FIELD_TYPE *s2));
 
 /* s1=c*s2 c complex */
-_DECLARE(void, spinor_field_mulc, (_SPINOR_FIELD_TYPE*, _COMPLEX c, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_mulc, (_SPINOR_FIELD_TYPE *s1, _COMPLEX c, _SPINOR_FIELD_TYPE *s2));
 
 /* r=s1+s2 */
-_DECLARE(void, spinor_field_add, (_SPINOR_FIELD_TYPE *r, _SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_add, (_SPINOR_FIELD_TYPE *r, _SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* r=s1-s2 */
-_DECLARE(void, spinor_field_sub, (_SPINOR_FIELD_TYPE *r, _SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_sub, (_SPINOR_FIELD_TYPE *r, _SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* s1+=s2 */
-_DECLARE(void, spinor_field_add_assign, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_add_assign, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* s1-=s2 */
-_DECLARE(void, spinor_field_sub_assign, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_sub_assign, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* s1=0 */
-_DECLARE(void, spinor_field_zero, (_SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_zero, (_SPINOR_FIELD_TYPE *s1));
 
 /* s1=-s2 */
-_DECLARE(void, spinor_field_minus, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_minus, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* s1=r1*s2+r2*s3 */
-_DECLARE(void, spinor_field_lc, (_SPINOR_FIELD_TYPE*, _REAL r1, _SPINOR_FIELD_TYPE*, _REAL r2, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_lc, (_SPINOR_FIELD_TYPE *s1, _REAL r1, _SPINOR_FIELD_TYPE *s2, _REAL r2, _SPINOR_FIELD_TYPE *s3));
 
 /* s1+=r1*s2+r2*s3 */
-_DECLARE(void, spinor_field_lc_add_assign, (_SPINOR_FIELD_TYPE*, _REAL r1, _SPINOR_FIELD_TYPE*, _REAL r2, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_lc_add_assign, (_SPINOR_FIELD_TYPE *s1, _REAL r1, _SPINOR_FIELD_TYPE *s2, _REAL r2, _SPINOR_FIELD_TYPE *s3));
 
 /* s1=c1*s2+c2*s3 c1, c2 complex*/
-_DECLARE(void, spinor_field_clc, (_SPINOR_FIELD_TYPE*, _COMPLEX c1, _SPINOR_FIELD_TYPE*, _COMPLEX c2, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_clc, (_SPINOR_FIELD_TYPE *s1, _COMPLEX c1, _SPINOR_FIELD_TYPE *s2, _COMPLEX c2, _SPINOR_FIELD_TYPE *s3));
 
 /* s1+=c1*s2+c2*s3 c1, c2 complex*/
-_DECLARE(void, spinor_field_clc_add_assign, (_SPINOR_FIELD_TYPE*, _COMPLEX c1, _SPINOR_FIELD_TYPE*, _COMPLEX c2, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_clc_add_assign, (_SPINOR_FIELD_TYPE *s1, _COMPLEX c1, _SPINOR_FIELD_TYPE *s2, _COMPLEX c2, _SPINOR_FIELD_TYPE *s3));
 
 /* s1=g5*s2  */
-_DECLARE(void, spinor_field_g5, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_g5, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* s1=g5*s1  */
-_DECLARE(void, spinor_field_g5_assign, (_SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_g5_assign, (_SPINOR_FIELD_TYPE *s1));
 
 /* s1=g0*s2  */
-_DECLARE(void, spinor_field_g0, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_g0, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* s1=g1*s2  */
-_DECLARE(void, spinor_field_g1, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_g1, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* s1=g2*s2  */
-_DECLARE(void, spinor_field_g2, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_g2, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* s1=g3*s2  */
-_DECLARE(void, spinor_field_g3, (_SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_g3, (_SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
 /* tools per eva.c  */
-_DECLARE(void, spinor_field_lc1, (_REAL c1, _SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_lc1, (_REAL c1, _SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
-_DECLARE(void, spinor_field_lc2, (_REAL c1, _REAL c2, _SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_lc2, (_REAL c1, _REAL c2, _SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2));
 
-_DECLARE(void, spinor_field_lc3, (_REAL c1, _REAL c2, _SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*, _SPINOR_FIELD_TYPE*));
+_DECLARE(void, spinor_field_lc3, (_REAL c1, _REAL c2, _SPINOR_FIELD_TYPE *s1, _SPINOR_FIELD_TYPE *s2, _SPINOR_FIELD_TYPE *s3));
 
 
 //undefine macros
