+++ Include/Update/TMPL/dirac.h.tmpl
@@ -58,6 +58,11 @@ _DECLARE(void, g5Dphi_eopre_sq, (double m0, _SPINOR_FIELD_TYPE *out, _SPINOR_FIE
 #undef _PTR_FUNC_INNER
 #undef _DECLARE
 
+#ifdef __cplusplus
+//there is no 'restrict' in C++, use an extension
+#undef restrict __restrict__
+#endif
+
 //user interface macros
 #undef _SPINOR_FIELD_TYPE
 #undef _SUFFIX
