+++ LibHR/Observables/measure_ff.c
@@ -9,6 +9,7 @@
 #include "io.h"
 #include "memory.h"
 #include "utils.h"
+#include "inverters.h"
 
 #define PI 3.141592653589793238462643383279502884197
