+++ LibHR/Observables/measure_mesons.c
@@ -9,6 +9,7 @@
 #include "io.h"
 #include "memory.h"
 #include "utils.h"
+#include "inverters.h"
 #include <string.h>
 
 #define PI 3.141592653589793238462643383279502884197
