+++ LibHR/Observables/measure_scattering.c
@@ -3,6 +3,7 @@
 #include "io.h"
 #include "memory.h"
 #include "utils.h"
+#include "inverters.h"
 
 #define PI 3.141592653589793238462643383279502884197
