+++ LibHR/Observables/wilsonloops.c
@@ -5,6 +5,7 @@
 #include "io.h"
 #include <string.h>
 #include "geometry.h"
+#include "inverters.h"
 
 #if defined(ROTATED_SF) && defined(BASIC_SF)
 #error This code does not work with the Schroedinger functional
