+++ LibHR/Observables/z2semwall_new.c
@@ -9,9 +9,9 @@
 #include "libhr_core.h"
 #include "memory.h"
 #include "io.h"
-#include "Inverters/linear_algebra.h"
 #include "random.h"
 #include "Update/dirac.h"
+#include "inverters.h"
 
 
 /* #define RESTRICTED_SET */
